mcp_servers/python/servers/DATAROBOT/mcp-datarobot/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-datarobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.4",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-datarobot = "mcp_datarobot.__main__:main"

[tool.uv]
package = true
